working etimers. The clock rate is slow and the seconds are meaningless, but it's going through the process properly.

+++ platform/mc1322x/contiki-conf.h
@@ -110,7 +110,7 @@ typedef int32_t s32_t;
 /* #define UIP_CONF_TCP_SPLIT       1 */
 
 
-typedef unsigned short clock_time_t;
+typedef uint32_t clock_time_t;
 typedef unsigned char u8_t;
 typedef unsigned short u16_t;
 typedef unsigned long u32_t;
